+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.12.3 (January 04,2023) 
+
+BUG FIXES:
+
+* resource/harness_platform_connector_github: Fix connector delete context. ([#377](https://github.com/harness/terraform-provider-harness/issues/377))
+
+# 0.12.2 (January 03,2023) 
+
+ENHANCEMENTS:
+
+* resource/harness_platform_usergroup: Add example to create user group by adding email. ([#373](https://github.com/harness/terraform-provider-harness/issues/373))
+
+# 0.12.1 (January 03,2023) 
+
+ENHANCEMENTS:
+
+* resource/harness_platform_usergroup:  Allow TF resource to support creating of user-groups by adding User email id. ([#371](https://github.com/harness/terraform-provider-harness/issues/371))
+
 # 0.12.0 (December 23,2022) 
 
 FEATURES: